requirements/common.txt
# Formatting, Linting, and Testing Dependencies
pytest==7.1.3
pytest-cov==3.0.0
pytest-playwright==0.4.4
jupyter==1.0.0
mypy==1.8.0
ruff==0.2.1
mock==5.1.0

# Mypy typing Dependencies
-r mypy-requirements.txt

# Requests Dependencies
requests==2.32.3